--- src/apis/formConfigAPI.js
@@ -1,11 +1,0 @@
-import { GET } from 'apis/http';
-
-/**
- * fetchFormConfig method fetch file data
- * @param {Object} fileDetails
- * @return {Promise}
- */
-export const fetchFormConfig = ({ tenant, fileName }) => (
-  GET({
-    url: `src/config/tenant/${tenant}/${fileName}.json`,
-  }));
